Refactor data source connection table
