+++ src/plugins/workspace/public/components/workspace_form/data_source_connection_table.scss
